Configured the addition of transport

+++ server/src/main/java/by/av/test/testavby/repository/transport/TransportBrandRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TransportBrandRepository extends JpaRepository<TransportBrand, Long> {
+public interface TransportBrandRepository extends JpaRepository<TransportBrand, Integer> {
     Optional<TransportBrand> findByBrandName(String brandName);
 }
